+++ CHANGELOG.md
@@ -8,6 +8,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 ## Fixed
 
 * Fixed `Konva.Text` justify drawing for a text with decoration
+* Added methods `data()`,`setData()`  and `getData()` methods to `Konva.TextPath`
 
 
 ## [2.1.3][2018-05-17]